initial-improvments
edit css.
make the syncro js as a function so you send it the id/class of the div and the data as json to render.
eliminate the foreach and to not reload all charts if multiple.
remove default data array.
add data and value to the segments and headers.
make the hover and click events for the chart to edit the size of the bar segment.
some improvments.